requirements.txt
awsebcli==3.14.3
blessed==1.15.0
boto3==1.7.58
botocore==1.10.58
cached-property==1.4.3
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
docker==3.4.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
Flask-JWT==0.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lml==0.0.1
lxml==4.2.3
MarkupSafe==1.0
numpy==1.14.5
pandas==0.23.3
pathspec==0.5.5
psycopg2==2.7.5
PyJWT==1.4.2
pynamodb==3.3.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.5
PyYAML==3.13
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
xlrd==1.1.0
